fix : z인덱스 오남용된 부분 삭제했습니다
