+++ src/1_Page/Championship/ui/DashBoard/ui/MatchListTab/model/type.d.ts
@@ -1,5 +1,6 @@
 type UseSearchTeamHandlerReturn = {
   searchTerm: string;
+  myMatchList: ChampionshipMatchList[];
   filteredMatches: ChampionshipMatchList[];
   handleSearchChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
 };
